refactor: change name from generator to intelligence
